fix: resolve all CodeQL error-severity alerts
- lemon.c: escape %destructor in format strings (% was parsed as %d specifier)
- dehacked.cpp: add missing num arg to error() format string
- m_argv.cpp: replace fixed char[256] buf with std::string to prevent overflow
- command.cpp: use memcpy with explicit size instead of strcpy (bounds clarity)
- w_wad.cpp: fix %4c -> %.4s for char array format specifier
- parser.cpp: cast ptrdiff_t to int for %d format specifier
- m_misc.cpp: fix off-by-one in allocation (len -> len+1 for null terminator)

+++ src/util/parser.cpp
@@ -344,7 +344,7 @@ bool Parser::ParseCmd(const parsercmd_t *commands, char *base)
 
     if (commands->name == NULL)
     {
-        CONS_Printf("Unknown command '%s' before char %d!\n", cmd, s - ms);
+        CONS_Printf("Unknown command '%s' before char %d!\n", cmd, (int)(s - ms));
         return false; // not found
     }
